# sso-swagger.yaml
openapi: 3.0.0
info:
  title: EZLIFE Real Estate SSO API
  version: 1.0.0
  description: API documentation for EZLIFE Real Estate SSO Backend
servers:
  - url: http://localhost:8080/api/v1
    description: Default API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Permission:
      type: object
      properties:
        id:
          type: integer
          format: int11
        url:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /permissions/read:
    get:
      tags:
        - Permissions
      summary: Get paginated list of permissions
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          required: true
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10
          required: true
          description: Number of items per page
      responses:
        '200':
          description: Successfully retrieved permissions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: 1
                  EM:
                    type: string
                    example: "Get list permissions at page 1, limit 10"
                  DT:
                    type: object
                    properties:
                      totalRows:
                        type: integer
                        example: integer
                      totalPages:
                        type: integer
                        example: integer
                      permissions:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            url:
                              type: string
                            description:
                              type: string
                        example:
                          - id: integer
                            url: string
                            description: string
                          - id: integer
                            url: string
                            description: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: -1
                  EM:
                    type: string
                    example: "You don't have permission to access this resource!"
                  DT:
                    type: string
                    example: ""   
  /permissions/create:
    post:
      tags:
        - Permissions
      summary: Create a new permission
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - url
                - description
              properties:
                url:
                  type: string

                description:
                  type: string

      responses:
        '201':
          description: Permission created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: 1
                  EM:
                    type: string
                    example: "Create permission successfully"
                  DT:
                    $ref: '#/components/schemas/Permission'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: -1
                  EM:
                    type: string
                    example: "You don't have permission to access this resource!"
                  DT:
                    type: string
                    example: ""  
  /permissions/assign-to-role:
    post:
      tags:
        - Permissions
      summary: Assign permissions to a role
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: object
                  required:
                    - roleId
                    - rolePermissions
                  properties:
                    roleId:
                      type: integer
                      example: integer
                    rolePermissions:
                      type: array
                      items:
                        type: object
                        required:
                          - permissionId
                        properties:
                          permissionId:
                            type: integer
                      example:
                        - permissionId: integer
                        - permissionId: integer
                        - permissionId: integer
                        - permissionId: integer
                        - permissionId: integer
                        - permissionId: integer
      responses:
        '200':
          description: Permission assigned successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: -1
                  EM:
                    type: string
                    example: "You don't have permission to access this resource!"
                  DT:
                    type: string
                    example: ""  
  /permissions/update/{id}:
    put:
      tags:
        - Permissions
      summary: Update permission by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission updated successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: -1
                  EM:
                    type: string
                    example: "You don't have permission to access this resource!"
                  DT:
                    type: string
                    example: ""  
  /permissions/delete/{id}:
    delete:
      tags:
        - Permissions
      summary: Delete permission by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission deleted successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  EC:
                    type: integer
                    example: -1
                  EM:
                    type: string
                    example: "You don't have permission to access this resource!"
                  DT:
                    type: string
                    example: ""  
...